# Conflicts:
#	demo/demo.cpp
#	demo/demo_frame_graphs.hpp
#	imgui.ini
#	src/render_tasks/d3d12_deferred_composition.cpp
#	src/render_tasks/d3d12_rt_hybrid_task.hpp

+++ src/d3d12/d3d12_renderer.cpp
@@ -1097,7 +1097,7 @@ namespace wr
 			LOGC("Nice try boiii! That's not a shape.");
 		}
 
-		return m_simple_shapes[type];
+		return m_simple_shapes[static_cast<std::size_t>(type)];
 	}
 
 	void D3D12RenderSystem::ResetBatches(SceneGraph & sg)
@@ -1155,7 +1155,7 @@ namespace wr
 				{ -1, 1, 1,		0, 0,		0, 1, 0,		0, 0, 0,	0, 0, 0,	0, 0, 0  },
 			};
 
-			m_simple_shapes[SimpleShapes::CUBE] = m_shapes_pool->LoadCustom<wr::VertexColor>({ mesh });
+			m_simple_shapes[static_cast<std::size_t>(SimpleShapes::CUBE)] = m_shapes_pool->LoadCustom<wr::VertexColor>({ mesh });
 		}
 
 		{
@@ -1173,7 +1173,7 @@ namespace wr
 				{ -1,  1,  0,		0, 1,		0, 0, -1,			0, 0, 1,		0, 1, 0,		0, 0, 0 },
 			};
 
-			m_simple_shapes[SimpleShapes::PLANE] = m_shapes_pool->LoadCustom<wr::VertexColor>({ mesh });
+			m_simple_shapes[static_cast<std::size_t>(SimpleShapes::PLANE)] = m_shapes_pool->LoadCustom<wr::VertexColor>({ mesh });
 		}
 	}
